pyproject.toml
[project]
name = "sample-slicer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typed-ffmpeg",
]

[project.optional-dependencies]
dev = [
    "basedpyright",
]

[tool.basedpyright]
typeCheckingMode = "recommended"
reportUnusedCallResult = false
reportAny = false